+++ src/language-yaml/index.js
@@ -14,6 +14,7 @@ const languages = [
     filenames: [
       ...data.filenames.filter((filename) => filename !== "yarn.lock"),
       ".prettierrc",
+      ".stylelintrc",
     ],
   })),
 ];