+++ tests/format/jsx/last-line/jsfmt.spec.js
@@ -1,2 +1,2 @@
-run_spec(__dirname, ["flow", "typescript"], { jsxBracketSameLine: true });
-run_spec(__dirname, ["flow", "typescript"], { jsxBracketSameLine: false });
+run_spec(__dirname, ["flow", "typescript"], { bracketSameLine: true });
+run_spec(__dirname, ["flow", "typescript"], { bracketSameLine: false });